requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
alpino-query==2.1.3
    # via -r requirements.in
asgiref==3.5.2
    # via django
basexclient==8.4.4
    # via -r requirements.in
django==4.0.6
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.13.1
    # via -r requirements.in
lxml==4.7.1
    # via alpino-query
pytz==2022.1
    # via djangorestframework
sqlparse==0.4.2
    # via django
